+++ src/r3_port_std/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+- Added the Windows backend.
+
 ## [0.1.3] - 2021-10-29
 
 This release only includes minor changes to the documentation and internal implementation.